refactor: move scenarios mapping to "compose" folder
