+++ src/index.ts
@@ -1,4 +1,5 @@
-export { createScenario as scenario, mapScenarios } from "@modules/scenarios";
+export { mapScenarios } from "@compose/map-scenarios";
+export { createScenario as scenario } from "@modules/scenarios";
 export { createReusableTest as reusableTest } from "@modules/reusable-tests";
 export { listeners } from "@modules/listeners";
 export { expect } from "@core/expectations";
