--- src/modules/reporting/index.ts
@@ -1,2 +1,0 @@
-export { ConsoleReporter } from "./console-reporter";
-export { report } from "./report";
