+++ vite.config.ts
@@ -20,6 +20,7 @@ export default defineConfig({
     ],
     resolve: {
         alias: {
+            "@compose": resolve(__dirname, "src/compose"),
             "@modules": resolve(__dirname, "src/modules"),
             "@core": resolve(__dirname, "src/core"),
             "@shared": resolve(__dirname, "src/shared")
